Rough Design Complete
Finished with the alpha look of the site. No details added like
background images, database pulling, music, etc. Just focusing on the
bare-bone looks of the site. Looking great so far.  Will work on
modularity for the next few updates to ensure that the cells won't
collapse on each other (even on phones and tablets).